backend/requirements.txt
Django==4.0.6
djangorestframework
django-filter
flake8
pytest
isort
psycopg2
pillow
python-dotenv
drf_extra_fields
gunicorn==19.9.0